Combine `DataType` and `LongNames` into a single `DataType` trait.
With this change, the user can just specify a register bitfield type, rather than having to specify both the primitive type and the long name. This cuts down on duplication. This also simplifies how aliased registers are specified, making the user manually write out the `Aliased<>` type.

+++ libraries/tock-register-interface/tests/peripheral.rs
@@ -10,7 +10,7 @@
 ::tock_registers::peripheral! {
     #[cfg(any())]
     CfgNoCompile {
-        0x0 => should_not_compile: u8(DoesNotExist) {},
+        0x0 => should_not_compile: DoesNotExist {},
     }
 }
 
@@ -18,7 +18,7 @@
 ::tock_registers::peripheral! {
     CfgNoField {
         #[cfg(any())]
-        0x0 => should_not_compile: u8(DoesNotExist) {},
+        0x0 => should_not_compile: DoesNotExist {},
     }
 }
